Refactor imports to use global constants instead of local constants
