+++ sportsphere-app/components/CalendarInput.jsx
@@ -2,7 +2,7 @@ import React, {useState} from 'react'
 import { View, TextInput, StyleSheet } from 'react-native'
 import DateTimePicker from '@react-native-community/datetimepicker';
 import { format } from 'date-fns';
-import { COLORS, ROUNDED, FONTSIZE } from '../constants'
+import { COLORS, ROUNDED, FONTSIZE } from '../global'
 
 // CalendarInput component
 export default function CalendarInput({date, setDate, datePicker, datePickerHandler}) {
@@ -49,8 +49,8 @@ export const styles = StyleSheet.create({
     borderColor: COLORS.inputBorder,
     padding: 5,
     backgroundColor: COLORS.inputArea,
-    borderRadius: ROUNDED.small,
-    fontSize: FONTSIZE.medium,
+    borderRadius: ROUNDED.medium,
+    fontSize: FONTSIZE.body,
     color: COLORS.text,
   },
 })
