+++ sportsphere-app/screens/ReminderScreen.jsx
@@ -1,6 +1,6 @@
 import { Text, StyleSheet, View } from 'react-native'
 import React, { Component } from 'react'
-import { COLORS } from '../constants'
+import { COLORS } from '../global'
 import ReminderItemList from '../components/ReminderItemList'
 
 export default class ReminderScreen extends Component {
